Add refactoring of styling
