+++ apps/innbyggerpanelet/src/app/app.module.scss
@@ -3,10 +3,5 @@
 /* VIEWS */
 
 /* COMPONENTS */
-@import './components/candidateProfile/candidateExperiencePoints';
-@import './components/misc/propertieValueField/propertieValueField';
 
 /* CONTAINERS */
-@import './containers/candidateProfile/candidateProfile';
-@import './containers/candidateProfile/candidateProperties';
-@import './containers/candidateProfile/candidatePerformedInsightWork';
