+++ apps/innbyggerpanelet/src/app/components/candidateProfile/CandidateExperiencePoints.module.scss
@@ -1,10 +1,10 @@
-.experience-points-container {
+.experiencePointsContainer {
     display: flex;
     flex-direction: row;
     margin-top: 2rem;
 }
 
-.experience-bar {
+.experienceBar {
     background-color: #99DEAD;
     width: 14rem;
     height: 0.9rem;
@@ -12,14 +12,15 @@
     border-bottom-left-radius: 0.5rem;
 }
 
-.no-experience-bar {
+.noExperienceBar {
     background-color: gray;
     width: 10rem;
     height: 0.9rem;
     border-top-right-radius: 0.5rem;
     border-bottom-right-radius: 0.5rem;
 }
-.experience-points-info {
+
+.experiencePointsInfo {
     display: flex;
     justify-content: space-between;
     margin-bottom: 2rem;
