--- apps/innbyggerpanelet/src/app/components/misc/propertieValueField/_propertieValueField.scss
@@ -1,14 +1,0 @@
-.propertie-value-field {
-    display: grid;
-    grid-template-columns: 50% 50%;
-    background-color: lightgray;
-    margin-top: 1rem;
-    margin-bottom: 1rem;
-
-    h1.navds-heading {
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        padding-left: 1rem;
-    }
-}
