+++ apps/innbyggerpanelet/src/app/containers/candidateProfile/CandidatePerformedInsight.module.scss
@@ -1,0 +1,3 @@
+.candidatePerformedInsigth {
+    margin-top: 1.5rem;
+}
