+++ apps/innbyggerpanelet/src/app/containers/candidateProfile/CandidateProfile.module.scss
@@ -1,11 +1,11 @@
-.candidate-profile {
+.candidateProfile {
     display: flex;
     flex-direction: column;
     width: 100%;
     max-width: 45rem;
 }
 
-.candidate-info {
+.candidateInfo {
     display: flex;
     flex-direction: column;
     height: 100%;
