+++ apps/innbyggerpanelet/src/app/containers/candidateProfile/CandidateProperties.module.scss
@@ -1,3 +1,3 @@
-.candidate-properties {
+.candidateProperties {
     margin-top: 1.5rem;
 }
