+++ apps/innbyggerpanelet/src/app/containers/candidateProfile/CandidateProperties.tsx
@@ -3,10 +3,12 @@ import React, { ReactElement } from 'react'
 import PropertieValueField from '../../components/misc/propertieValueField/PropertieValueField'
 import { Candidate } from './CandidateProfile'
 
+import style from './CandidateProperties.module.scss'
+
 export default function CandidateProperties(candidate: Candidate): ReactElement {
     return (
         <div>
-            <div className='candidate-properties'>
+            <div className={style.candidateProperties}>
                 <Heading size="large">Egenskaper</Heading>
                 <PropertieValueField 
                     propertie="Alder"
