--- apps/innbyggerpanelet/src/app/containers/candidateProfile/_candidatePerformedInsightWork.scss
@@ -1,3 +1,0 @@
-.candidate-performed-insigth-work {
-    margin-top: 1.5rem;
-}
